// src/api/stetsom/model/publicVariant.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { PublicVariantAttr } from "./publicVariantAttr";

export interface PublicVariant {
  variant_id: string;
  /** @maxLength 200 */
  name: string;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  order: number;
  attributes: PublicVariantAttr[];
}